Returning errors from csharpier cli. Make sure we don't hang the old stdout version when cli fails
