+++ Src/CSharpier.Cli/PublicAPI.Shipped.txt
@@ -1,0 +1,1 @@
+﻿#nullable enable
