+++ Src/CSharpier.Cli/Server/FormatFileParameter.cs
@@ -1,0 +1,7 @@
+namespace CSharpier.Cli.Server;
+
+public class FormatFileParameter
+{
+    public required string fileContents { get; set; }
+    public required string fileName { get; set; }
+}
