+++ Src/CSharpier.Cli/Server/FormatFileResult.cs
@@ -1,0 +1,12 @@
+namespace CSharpier.Cli.Server;
+
+using System.Text.Json.Serialization;
+
+public class FormatFileResult(Status status)
+{
+    public string? formattedFile { get; set; }
+
+    [JsonConverter(typeof(JsonStringEnumConverter))]
+    public Status status { get; set; } = status;
+    public string? errorMessage { get; set; }
+}
