+++ Src/CSharpier.Cli/Server/Status.cs
@@ -1,0 +1,8 @@
+namespace CSharpier.Cli.Server;
+
+public enum Status
+{
+    Formatted,
+    Ignored,
+    Failed
+}
